requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-jose==3.5.0
email-validator==2.2.0
python-multipart==0.0.20
pydub==0.25.1
# audioop-lts==0.1.0